feat: Fix unittests
cleanup console test

Issue #9
